Move data classes to a separate module
